Update zip_choropleth.R
This change adds an optional "palette" variable to the zip_choropleth function so that the user has the ability to set the color palette for the Alaska and Hawaii insets.  Currently if you try to set the color via scale_fill_brewer it does not update the Alaska and Hawaii colors.  Example below, using data(df_pop_zip):
df_pop_zip %>% 
  zip_choropleth2(num_colors = 8,title = "Testing") +
  scale_fill_brewer(palette="RdYlGn", drop=FALSE)

Change will allow you to do the following:
df_pop_zip %>% 
  zip_choropleth2(num_colors = 8,title = "Testing",palette = "RdYlGn")

Open for other suggestions but I am needing to apply a custom color palette.